Split files to separate dmx class
